requirements.txt
flask
mysql-connector-python
Flask>=2.0
Flask-SQLAlchemy>=3.0
pymysql
Werkzeug
twilio
reportlab